Добавил тестирование для реализации методов работы с БД

+++ src/main/java/dao/TaskDAO.java
@@ -11,7 +11,7 @@ public interface TaskDAO {
     void delete(Long number);
     void update(Task task);
     List<Task> get(Date fromDate, Date toDate);
-    List<Task> get(Date fromDate, Date toDate, Long number);
+    List<Task> get(Date fromDate, Date toDate, Integer number);
     Task getByNumber(Long number);
 
 }
